requirements.txt
asgiref==3.6.0
Django==4.1.6
djangorestframework==3.14.0
pip==23.0
psycopg2-binary==2.9.5
pytz==2022.7.1
setuptools==67.0.0
sqlparse==0.4.3
wheel==0.38.4
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
python-decouple==3.7
email_validator==1.3.1
django-filter==22.1